After some tests, we need new ideas.
I add a comparison with the regular FD method of 4th order (centered
5-point stencil). Calculating the divergent at the origin we get
something of the order of 10^-7, while using divFreeFD we get 10^-11.

What would be really great is to get the 4th order decay without adding
points to the stencil. I'm not sure, but maybe polynomial interpolation
can be useful to introduce this extra interpolation points.