I simplified the code of the stream function. The code is easier to read now. I also add 2 files contaning the interpolation points for u and v. 1 stands for use as interpolation point and 0 to not use it.
I did some tests choosing different interpolation points but I wasn't able
to increase the accuracy of the derivatives not related to the divergence
free equation.

In other hand, now we can get a linear system that is not rank deficient
and still get the double convergence on the derivatives related to the
divergence free equation.