Second derivative can not get 4th order accuracy using polynomials of degree 3 (N = 3). If using degree 4 we can get 4th order accuracy on uxx and vyy.
I add a file call gridForUandV. This file show the grid points of U
and V if we use different spacements for x and y direction, for each
grid. The idea is to use the double of points of the x direction on
the y direction for getting the error about the same order for ux and
uy. It is similar for the grid of v, but we switch the directions.