Add a file for visualization of Lebesgue function
The file lebesgueFunctions.m plots the cardinal and lebesgue
function. Change the uInterpPts.txt and vInterpPts.txt to see how
the interpolation points change the Lebesgue functions and the
Lebesgue constant.

As future work, we should compare those cardinal and Lebesgue
functions with the ones from the divergence free RBF method.