Looking to the limit using chebfun2
Add functions to use chebfun2 with comparingSymAndCP.m script. The idea is
to use the limit from the countour-Padè algorithm (Cauchy integral) and
then use the chebfun2 to get a polynomial approximation.

New files:

chebpoly.m : creates the a symbolic expression for the Chebyshev
             polynomial.

chebpoly2function.m : truncates the chebfun2 approximation and output a
                      anonymous function.