Faster convergence in both derivatives
If we rotate and jitter the grid points we can get some faster convergence
for the derivatives uy and vx, but loose some convergence for ux and vy.
However, if we just rotate the grid points we loose the faster convergence
for the derivatives in the divergence-free condition and gain faster
convergence in other directions (the x and y axis rotated).

More test are necessary to understand why we get this faster convergence,
principally when we have some random distribution of points.